requirements.txt
numpy
tensorflow
matplotlib
pandas
Pillow